feat: move basic stuff to lib
