+++ src/server/driver/Fetcher/ICSFetcher.ts
@@ -1,4 +1,4 @@
-import AppError from "~/utils/error";
+import AppError from "~/lib/error/error";
 import type { Fetcher } from "./Fetcher";
 
 export class ICSFetcher implements Fetcher {
