+++ src/server/widgets/links/widget.tsx
@@ -1,4 +1,4 @@
-import { useDetectScreenSize } from "~/utils/hooks";
+import { useDetectScreenSize } from "~/lib/ui/hooks";
 import type { WidgetProps } from "~/utils/types/widget";
 import LinkItem from "./LinkItem";
 import type { LinkWidgetData } from "./types";
